requirements.txt
# Зависимости для ЛР2
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
Pillow>=9.0.0
onnx>=1.13.0
onnxscript>=0.1.0
